+++ src/model/show.rs
@@ -80,7 +80,10 @@ pub struct SimplifiedEpisode {
     pub images: Vec<Image>,
     pub is_externally_hosted: bool,
     pub is_playable: bool,
-    /// Note: This field is deprecated and might be removed in the future. Please use the languages field instead
+    #[deprecated(
+        since = "0.11",
+        note = "This `language` field is deprecated and might be removed in the future by Spotify. Please use the languages field instead"
+    )]
     pub language: String,
     pub languages: Vec<String>,
     pub name: String,